arm64: dts: 19805/20809/20828: Don't blank screen as part of LP1 cmd
We send the LP1 command as part of entering Always-On Display mode, but
we don't want that command to blank the screen. After all, the point of
AoD mode is to keep the screen on to display the time, notifications,
etc.

This change was previously done for 19821/19855 (OnePlus 8), but the
same issue affects 19805/20809/20828 (OnePlus 8T and OnePlus 9R) too.

Change-Id: Ica1a5da4441ba7e9208448f58cb7f4ced2c662a1
Signed-off-by: Pajdek <pajdek@gmail.com>